requirements.txt
plotly==5.6.0
plotly-express==0.4.1
pandas==1.4.2
numpy==1.21.5
folium==0.14.0
matplotlib==3.5.1
matplotlib-inline==0.1.2
haversine==2.8.0
streamlit-folium==0.11.1
inflection==0.5.1
seaborn==0.11.2
wordcloud==1.9.2
stopwords==1.0.0
streamlit==1.23.1
pillow=9.0.1